Safari-609.1.11.1/LayoutTests/js/dfg-arguments-mixed-alias-expected.txt

Tests functions that use 'arguments' in both an aliased and a non-aliased way.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bar(42) is 42 on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
